requirements-free.txt
# Minimal dependencies for free version
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
motor==3.3.2
beanie==1.23.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
python-jobspy==1.1.50
